+++ proj/code/include/ray.hpp
@@ -5,42 +5,45 @@
 #include <iostream>
 #include <Vector3f.h>
 
-
 // Ray class mostly copied from Peter Shirley and Keith Morley
-class Ray {
+class Ray
+{
 public:
-
     Ray() = delete;
-    Ray(const Vector3f &orig, const Vector3f &dir) {
+    Ray(const Vector3f &orig, const Vector3f &dir)
+    {
         origin = orig;
         direction = dir;
     }
 
-    Ray(const Ray &r) {
+    Ray(const Ray &r)
+    {
         origin = r.origin;
         direction = r.direction;
     }
 
-    const Vector3f &getOrigin() const {
+    const Vector3f &getOrigin() const
+    {
         return origin;
     }
 
-    const Vector3f &getDirection() const {
+    const Vector3f &getDirection() const
+    {
         return direction;
     }
 
-    Vector3f pointAtParameter(float t) const {
+    Vector3f pointAtParameter(double t) const
+    {
         return origin + direction * t;
     }
 
 private:
-
     Vector3f origin;
     Vector3f direction;
-
 };
 
-inline std::ostream &operator<<(std::ostream &os, const Ray &r) {
+inline std::ostream &operator<<(std::ostream &os, const Ray &r)
+{
     os << "Ray <" << r.getOrigin() << ", " << r.getDirection() << ">";
     return os;
 }
